+++ src/freak_media_player/plugins/visualizer/__init__.py
@@ -1,0 +1,5 @@
+"""Built-in audio visualizer plugin."""
+
+from freak_media_player.plugins.visualizer.plugin import VisualizerPlugin
+
+__all__ = ["VisualizerPlugin"]